Update zebra_datepicker.src.js
Hi @stefangabos,

I found your plugin pretty interesting, and I really like the implementation. I am currently using it, but I have had to tweak few things in order to have what I was looking for.

I just added a new option, and tweaked a bit your code in few points, in order to allow the plugin to setup an internal div container in each cell. I am using the datepicker with a complete different layout, and due to the huge size I am using, I have had the necessity to add this containers.
The flag option is used to set this behaviour or to just leave your plugin untouched.
I thought that this could be a nice add on for an already superb work.

Please looking forward to hearing from you (your opinions or comments are more than welcome).

Best
Dinuz