Aligning year selector to use zero indexed months
When using a year selector, the datepicker will return February of the selected year due to the fact that the JavaScript Date object's `month` value is zero index based. Updating the value passed here will direct the `select_date` function to pass the correct `month` value to the Date object.