Added an option "enabled_ampm"
This works like enabled_seconds, enabled_minutes, or enabled_hours.
It takes an array, and can be set to ['am'] only or ['pm'] only.
This is a necessary option when setting minimum/maximum times on a timepicker input that is in 12 hr format.

Has been tested and works. Feel free to test more extensively.
zebra_datepicker.min.js will need to be updated to include this new code, and you may want to change my syntax a bit, as I noticed you don't use as many brackets as I do in your code.